Update frontend + authorization login and access

+++ app/Http/Controllers/HobbyController.php
@@ -6,6 +6,8 @@ use App\Models\Hobby;
 use App\Models\Tag;
 use Illuminate\Http\Request;
 use Illuminate\Support\Facades\Session;
+use Intervention\Image\Facades\Image;
+use Illuminate\Support\Facades\Gate;
 
 
 class HobbyController extends Controller
@@ -60,13 +62,11 @@ class HobbyController extends Controller
             'user_id' => auth()->id()
         ]);
         $hobby->save();
-        /*
-   return $this->index()->with(	        return $this->index()->with(
-       [	            [
-           'message_success' => "The hobby <b>" . $hobby->name . "</b> was created."	                'message_success' => "The hobby <b>" . $hobby->name . "</b> was created."
-       ]	            ]
-   );	        );
-   */
+
+        if($request->image){
+            $this->saveImages($request->image, $hobby->id);
+        }
+
         return redirect('/hobby/' . $hobby->id)->with(
             [
                 'message_warning' => "Please assign some tags now."
@@ -102,8 +102,12 @@ class HobbyController extends Controller
      */
     public function edit(Hobby $hobby)
     {
+        abort_unless(Gate::allows('update', $hobby), 403);
+
         return view('hobby.edit')->with([
-            'hobby' => $hobby
+            'hobby' => $hobby,
+            'message_success' => Session::get('message_success'),
+            'message_warning' => Session::get('message_warning')
         ]);
     }
 
@@ -116,10 +120,17 @@ class HobbyController extends Controller
      */
     public function update(Request $request, Hobby $hobby)
     {
+        abort_unless(Gate::allows('update', $hobby), 403);
+
         $request->validate([
             'name' => 'required|min:3',
             'description' => 'required|min:5',
+            'image' => 'mimes:jpeg,jpg,bmp,png,gif'
         ]);
+
+        if($request->image){
+            $this->saveImages($request->image, $hobby->id);
+        }
 
         $hobby->update([
             'name' => $request['name'],
@@ -139,11 +150,52 @@ class HobbyController extends Controller
      */
     public function destroy(Hobby $hobby)
     {
+        abort_unless(Gate::allows('delete', $hobby), 403);
+
         $oldName = $hobby->name;
         $hobby->delete();
 
         return $this->index()->with([
             'message_success' => 'The hobby  <b>' .  $oldName . "</b> was deleted."
         ]);
+    }
+
+    public function saveImages($imageInput, $hobby_id){
+        $image = Image::make($imageInput);
+
+        if($image -> width() > $image -> height()) { //landscape
+            $image->widen(1200)
+                ->save(public_path() . "/img/hobbies/" . $hobby_id . "_large.jpg")
+                ->widen(400) -> pixelate(12)
+                ->save(public_path() . "/img/hobbies/" . $hobby_id . "_pixelated.jpg");
+            $image = Image::make($imageInput);
+            $image->widen(60)
+                ->save(public_path() . "/img/hobbies/" . $hobby_id . "_thumb.jpg");
+        }
+        else{ //portrait
+            $image->heighten(900)
+                ->save(public_path() . "/img/hobbies/" . $hobby_id . "_large.jpg")
+                ->heighten(400) -> pixelate(12)
+                ->save(public_path() . "/img/hobbies/" . $hobby_id . "_pixelated.jpg");
+            $image = Image::make($imageInput);
+            $image->heighten(60)
+                ->save(public_path() . "/img/hobbies/" . $hobby_id . "_thumb.jpg");
+        }
+    }
+
+    public function deleteImages($hobby_id){
+        if(file_exists(public_path() . "/img/hobbies/" . $hobby_id . "_large.jpg"))
+            unlink(public_path() . "/img/hobbies/" . $hobby_id . "_large.jpg");
+        if(file_exists(public_path() . "/img/hobbies/" . $hobby_id . "_pixelated.jpg"))
+            unlink(public_path() . "/img/hobbies/" . $hobby_id . "_pixelated.jpg");
+        if(file_exists(public_path() . "/img/hobbies/" . $hobby_id . "_thumb.jpg"))
+            unlink(public_path() . "/img/hobbies/" . $hobby_id . "_thumb.jpg");
+
+        return back()->with(
+            [
+                'message_success' => "The Image was deleted."
+            ]
+        );
+
     }
 }
